HVDCWISE_TEA

Documentation for HVDCWISE_TEA.

HVDCWISE_TEA._DCdata

HVDCWISE_TEA._UNUSED_VARS

HVDCWISE_TEA._conductor_matrix

HVDCWISE_TEA._conductorless

HVDCWISE_TEA.add_slack_operation_cost!

HVDCWISE_TEA.add_slack_operation_cost!

HVDCWISE_TEA.conductorsDC_status

HVDCWISE_TEA.constraint_flexible_demand

HVDCWISE_TEA.constraint_flexible_demand

HVDCWISE_TEA.constraint_power_balance_ac

HVDCWISE_TEA.constraint_power_balance_ac

HVDCWISE_TEA.make_multiconductor!

HVDCWISE_TEA.make_multinetwork

HVDCWISE_TEA.parse_data

HVDCWISE_TEA.parse_file

HVDCWISE_TEA.process_additional_data!

HVDCWISE_TEA.ref_add_flex_load!

HVDCWISE_TEA.terminalDC_voltage_bound

HVDCWISE_TEA.terminalDC_voltage_start

HVDCWISE_TEA.variable_demand_reduction

HVDCWISE_TEA.variable_demand_shifting_downwards

HVDCWISE_TEA.variable_demand_shifting_upwards

HVDCWISE_TEA.variable_flexible_demand

HVDCWISE_TEA.variable_mcdc_converter

HVDCWISE_TEA.variable_mcdc_converter

HVDCWISE_TEA.variable_slack_power

HVDCWISE_TEA.variable_slack_power

HVDCWISE_TEA.variable_slack_power_imaginary

HVDCWISE_TEA.variable_slack_power_real

HVDCWISE_TEA.variable_total_demand

HVDCWISE_TEA._DCdata — Constant

only dc side data

HVDCWISE_TEA._UNUSED_VARS — Constant

list of all PowerModelsunused properties inHVDCWISE_TEA` optimisation model

HVDCWISE_TEA._conductor_matrix — Constant

field names that should become multiconductor matrix not arrays

HVDCWISE_TEA._conductorless — Constant

field names that should not be multiconductor values

HVDCWISE_TEA.add_slack_operation_cost! — Method

expression for nodal slack power cost

HVDCWISE_TEA.add_slack_operation_cost! — Method

expression for nodal slack power cost

HVDCWISE_TEA.conductorsDC_status — Method

Generate vector of multi-conductor status states for convdc and branchdc components

HVDCWISE_TEA.constraint_flexible_demand — Method

constraint on total active power demand accounting for all flexibility components

HVDCWISE_TEA.constraint_flexible_demand — Method

constraint on total flexible load power

HVDCWISE_TEA.constraint_power_balance_ac — Method

nodal power balance constraint of hybrid AC/DC multi-conductor network including storage & flexible demand

HVDCWISE_TEA.constraint_power_balance_ac — Method

Power balance of hybrid AC/DC network with storage & flexible demand

HVDCWISE_TEA.make_multiconductor! — Method

make_multiconductor!(data::Dict{String,<:Any}, components::Vector{String})

Generate a multinetwork data structure from a single network and a time series.

Arguments

data: single-conductor data structure to be expanded.

components: network DC components to which the multi-conductor transformation is applied. Default to list of all modelled DC component names, i.e. ["busdc", "convdc", "branchdc"].

HVDCWISE_TEA.make_multinetwork — Method

make_multinetwork(sn_data, time_series; <keyword arguments>)

Generate a multinetwork data structure from a single network and a time series.

Arguments

sn_data: single-network data structure to be replicated.

time_series: data structure containing the time series.

global_keys: keys that are stored once per multinetwork (they are not repeated in each nw).

number_of_nws: number of networks to be created from sn_data and time_series; default: read from dim.

nw_id_offset: optional value to be added to time_series ids to shift nw ids in multinetwork data structure; default: read from dim.

check_dim: whether to check for dim in sn_data; default: true.

HVDCWISE_TEA.parse_data — Method

parse_data(path2grid::String, path2data::String; kwargs...)

Wrapper function for parsing, sequentially:

A Matpower .m file located at path2grid into an HVDCWISE_TEA single-network data structure;

A nested folder structure located at path2data into a dictionary of timeseries;

The single-network structure into a multi-network one by merging it with the timeseries dictionary, if provided;

The single or multi-network structure into a multi-conductor one (for the HVDC grid part, if provided).

HVDCWISE_TEA.parse_file — Method

parse_file(file; <keyword arguments>)

Parse a Matpower .m file into an HVDCWISE_TEA single-network data structure.

Keyword arguments, if any, are forwarded to PowerModels.parse_file.

HVDCWISE_TEA.process_additional_data! — Method

process_additional_data!(data::Dict{String,Any})

Organise, process and perform checks on data tables that are not used/recognised in PowerModels. These may include non-dispatchable generators, flexible loads, storages, phase-shift transformers and DC components (i.e., AC/DC converters and DC branches).

Optional tables are:

gencost and storage, which are processed in PowerModels;

ndgen, storage_extra, load_extra, which are processed in FlexPlan;

busdc, convdc and branchdc, which are processed in PowerModelsMCDC;

pst, which is processed in CbaOPF.

Other tables may be added as well: they will be made available in the returned object, without any check.

By default, element properties that are not used in HVDCWISE_TEA mathematical model are removed.

HVDCWISE_TEA.ref_add_flex_load! — Method

Add to ref the keys for handling flexible demand

HVDCWISE_TEA.terminalDC_voltage_bound — Method

Adjust voltage bound for multi-conductor busdc terminals

HVDCWISE_TEA.terminalDC_voltage_start — Method

Adjust voltage start value for multi-conductor busdc terminals

HVDCWISE_TEA.variable_demand_reduction — Method

Variable for power not consumed (voluntary load reduction) at each flexible load

HVDCWISE_TEA.variable_demand_shifting_downwards — Method

Variable for downward demand shifting at each flexible load

HVDCWISE_TEA.variable_demand_shifting_upwards — Method

Variable for upward demand shifting at each flexible load

HVDCWISE_TEA.variable_flexible_demand — Method

variables for flexible load

HVDCWISE_TEA.variable_mcdc_converter — Method

variables for AC/DC converters

HVDCWISE_TEA.variable_mcdc_converter — Method

variables for AC/DC converters

HVDCWISE_TEA.variable_slack_power — Method

variables for active and reactive bus slack

HVDCWISE_TEA.variable_slack_power — Method

variables for bus power slack

HVDCWISE_TEA.variable_slack_power_imaginary — Method

variable for bus reactive power slack

HVDCWISE_TEA.variable_slack_power_real — Method

variable for bus active power slack

HVDCWISE_TEA.variable_total_demand — Method

Variable for the actual active load demand at each load accounting for market flexibility