HVDCWISE_TEA

Documentation for HVDCWISE_TEA.

HVDCWISE_TEA.constraint_power_balance_ac

HVDCWISE_TEA.process_additional_data!

HVDCWISE_TEA.ref_add_flex_load!

HVDCWISE_TEA.constraint_power_balance_ac — Method

Power balance of hybrid AC/DC network with storage & flexible demand

HVDCWISE_TEA.process_additional_data! — Method

process_additional_data!(data::Dict{String,Any})

Organise, process and perform checks on data tables that are not used/recognised in PowerModels. These may include non-dispatchable generators, storages, flexible loads, phase-shift transformers and DC components (i.e. AC/DC converters and DC branches).

Optional tables are:

gencost and storage, which are processed in PowerModels;

ngen, storage_extra, load_extra, which are processed in FlexPlan;

busdc, convdc and branchdc, which are processed in PowerModelsMCDC;

pst, which is processed in CbaOPF.

Other tables may be added as well: they will be made available in the returned object, without any check.

Extended help

The available load_extra table parameters are:

flex : boolean thta defines if load is flexible or not. If not flexible, only involutary curtailment is active.

pf_angle : fixed power factor angle in radians, giving the reactive power as Q = P ⨉ tan(θ)

cost_curt : compensation for involuntary load curtailment (i.e. involuntary demand reduction) [€/MWh];

cost_red : compensation for consuming less (i.e. voluntary demand reduction) [€/MWh];

cost_shift : Compensation for demand shifting, applied half to the power shifted upward and half to the power shifted downward [€/MWh]

co2_cost : CO2 costs for enabling flexible demand [€]

pred_rel_max : superior bound on voluntary load reduction (not consumed power) as a fraction of the total reference demand [p.u.];

pshift_up_rel_max : superior bound on upward demand shifted as a fraction of the total reference demand [p.u.];

pshift_down_rel_max: superior bound on downward demand shifted as a fraction of the total reference demand [p.u.];

ered_rel_max : superior bound on voluntary energy reduction as a fraction of the total reference demand [p.u.];

eshift_rel_max : superior bound on shifted energy as a fraction of the total reference demand [p.u.];

tshift_up : Recovery period for upward demand shifting [h];

tshift_down : Recovery period for downward demand shifting [h];

HVDCWISE_TEA.ref_add_flex_load! — Method

Add to ref the keys for handling flexible demand